+++ test/measurements/charge_test.dart
@@ -176,7 +176,7 @@ void main() {
       });
       test('maintains units', () {
         // given
-        final measurement = amperes(3.4).withPrecision(Precision(3));
+        final measurement = amperes(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -186,7 +186,7 @@ void main() {
       });
       test('maintains prefix', () {
         // given
-        final measurement = milli.amperes(3.4).withPrecision(Precision(3));
+        final measurement = milli.amperes(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -196,7 +196,7 @@ void main() {
       });
       test('extension maintains prefix', () {
         // given
-        final measurement = 3.4.milli.amperes.withPrecision(Precision(3));
+        final measurement = 3.4.milli.amperes.withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -206,10 +206,10 @@ void main() {
       });
       test('modified precision', () {
         // given
-        final measurement = deci.amperes(23.45).withPrecision(Precision(3));
+        final measurement = deci.amperes(23.45).withPrecisionOf(3);
 
         // when
-        final result = measurement.withPrecision(Precision(2)).toString();
+        final result = measurement.withPrecisionOf(2).toString();
 
         // then
         expect(result, '23.0 dA');