+++ test/measurements/distance_test.dart
@@ -297,7 +297,7 @@ void main() {
       });
       test('maintains units', () {
         // given
-        final measurement = feet(3.4).withPrecision(Precision(3));
+        final measurement = feet(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -307,7 +307,7 @@ void main() {
       });
       test('maintains prefix', () {
         // given
-        final measurement = milli.feet(3.4).withPrecision(Precision(3));
+        final measurement = milli.feet(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -317,7 +317,7 @@ void main() {
       });
       test('extension maintains prefix', () {
         // given
-        final measurement = 3.4.milli.feet.withPrecision(Precision(3));
+        final measurement = 3.4.milli.feet.withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -327,17 +327,17 @@ void main() {
       });
       test('modified precision', () {
         // given
-        final measurement = deci.feet(23.45).withPrecision(Precision(3));
+        final measurement = deci.feet(23.45).withPrecisionOf(3);
 
         // when
-        final result = measurement.withPrecision(Precision(2)).toString();
+        final result = measurement.withPrecisionOf(2).toString();
 
         // then
         expect(result, '23.0 dft');
       });
       test('modified units', () {
         // given
-        final measurement = deci.feet(23.45).withPrecision(Precision(3));
+        final measurement = deci.feet(23.45).withPrecisionOf(3);
 
         // when
         final result = measurement.withDefaultUnit(milli.meters).toString();