+++ test/measurements/luminosity_test.dart
@@ -221,7 +221,7 @@ void main() {
       });
       test('maintains units', () {
         // given
-        final measurement = candlepower(3.4).withPrecision(Precision(3));
+        final measurement = candlepower(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -231,7 +231,7 @@ void main() {
       });
       test('maintains prefix', () {
         // given
-        final measurement = milli.candlepower(3.4).withPrecision(Precision(3));
+        final measurement = milli.candlepower(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -241,7 +241,7 @@ void main() {
       });
       test('extension maintains prefix', () {
         // given
-        final measurement = 3.4.milli.candlepower.withPrecision(Precision(3));
+        final measurement = 3.4.milli.candlepower.withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -251,17 +251,17 @@ void main() {
       });
       test('modified precision', () {
         // given
-        final measurement = deci.candlepower(23.45).withPrecision(Precision(3));
+        final measurement = deci.candlepower(23.45).withPrecisionOf(3);
 
         // when
-        final result = measurement.withPrecision(Precision(2)).toString();
+        final result = measurement.withPrecisionOf(2).toString();
 
         // then
         expect(result, '23.0 dcp');
       });
       test('modified units', () {
         // given
-        final measurement = deci.candlepower(23.45).withPrecision(Precision(3));
+        final measurement = deci.candlepower(23.45).withPrecisionOf(3);
 
         // when
         final result =