+++ test/measurements/pressure_test.dart
@@ -110,7 +110,7 @@ void main() {
         ]);
 
         // when
-        final result = measurement.withPrecision(Precision(3)).as(pascals);
+        final result = measurement.withPrecisionOf(3).as(pascals);
 
         // then
         expect(result, 151000);
@@ -363,7 +363,7 @@ void main() {
       });
       test('maintains units', () {
         // given
-        final measurement = psi(3.4).withPrecision(Precision(3));
+        final measurement = psi(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -373,7 +373,7 @@ void main() {
       });
       test('maintains prefix', () {
         // given
-        final measurement = milli.pascals(3.4).withPrecision(Precision(3));
+        final measurement = milli.pascals(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -383,7 +383,7 @@ void main() {
       });
       test('extension maintains prefix', () {
         // given
-        final measurement = 3.4.milli.pascals.withPrecision(Precision(3));
+        final measurement = 3.4.milli.pascals.withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -393,17 +393,17 @@ void main() {
       });
       test('modified precision', () {
         // given
-        final measurement = deci.pascals(23.45).withPrecision(Precision(3));
+        final measurement = deci.pascals(23.45).withPrecisionOf(3);
 
         // when
-        final result = measurement.withPrecision(Precision(2)).toString();
+        final result = measurement.withPrecisionOf(2).toString();
 
         // then
         expect(result, '23.0 dPa');
       });
       test('modified units', () {
         // given
-        final measurement = deci.pascals(23.45).withPrecision(Precision(3));
+        final measurement = deci.pascals(23.45).withPrecisionOf(3);
 
         // when
         final result = measurement.withDefaultUnit(milli.bars).toString();