+++ test/measurements/temperature_change_test.dart
@@ -220,7 +220,7 @@ void main() {
       });
       test('maintains units', () {
         // given
-        final measurement = fahrenheit(3.4).withPrecision(Precision(3));
+        final measurement = fahrenheit(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -230,7 +230,7 @@ void main() {
       });
       test('maintains prefix', () {
         // given
-        final measurement = milli.fahrenheit(3.4).withPrecision(Precision(3));
+        final measurement = milli.fahrenheit(3.4).withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -240,7 +240,7 @@ void main() {
       });
       test('extension maintains prefix', () {
         // given
-        final measurement = 3.4.milli.fahrenheit.withPrecision(Precision(3));
+        final measurement = 3.4.milli.fahrenheit.withPrecisionOf(3);
 
         // when
         final result = measurement.toString();
@@ -250,17 +250,17 @@ void main() {
       });
       test('modified precision', () {
         // given
-        final measurement = deci.fahrenheit(23.45).withPrecision(Precision(3));
+        final measurement = deci.fahrenheit(23.45).withPrecisionOf(3);
 
         // when
-        final result = measurement.withPrecision(Precision(2)).toString();
+        final result = measurement.withPrecisionOf(2).toString();
 
         // then
         expect(result, '23.0 d°F');
       });
       test('modified units', () {
         // given
-        final measurement = deci.fahrenheit(23.45).withPrecision(Precision(3));
+        final measurement = deci.fahrenheit(23.45).withPrecisionOf(3);
 
         // when
         final result = measurement.withDefaultUnit(milli.celcius).toString();